feat: add tagsAnyOf and tagsAllOf filters for artworks
This makes it possible for users to filter for artworks that contain any of the tags listed, or contain all of the tags listed.
For this to work, I needed to refactor the search.ts stuff so external code can add its own specific "operator handlers", or "filter appliers", like we just did with artwork tags.

Also a few small fixes